backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hongyun-ledger-api"
version = "0.1.0"
description = "Backend API for 鸿运账本"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "fastapi>=0.115.12,<1.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "pyjwt>=2.10.1,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
